requirements.txt
torch
torchvision
pillow
flask